ECA-5629: Correct drop table SQL scripts for Oracle and HSQL. JavaDoc nameSpace.
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@25078 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ doc/sql-scripts/drop-tables-ejbca-hsqldb.sql
@@ -27,7 +27,7 @@ drop table PeerData if exists;
 drop table ProfileData if exists;
 drop table PublisherData if exists;
 drop table PublisherQueueData if exists;
-drop table RoleData;
+drop table RoleData if exists;
 drop table RoleMemberData if exists;
 drop table ServiceData if exists;
 drop table UserData if exists;
